+++ components/ui/Rating/Rating.tsx
@@ -7,7 +7,7 @@ export interface RatingProps {
   value: number
 }
 
-const Quantity: FC<RatingProps> = ({ value = 5 }) => {
+const Quantity: React.FC<RatingProps> = React.memo(({ value = 5 }) => {
   return (
     <div className="flex flex-row py-6 text-accent-9">
       {rangeMap(5, (i) => (
@@ -22,6 +22,6 @@ const Quantity: FC<RatingProps> = ({ value = 5 }) => {
       ))}
     </div>
   )
-}
+})
 
 export default Quantity
